+++ docs/fr/releases.rst
@@ -2,5 +2,5 @@ Notes de versions
 =================
 
 Les notes de versions sont disponibles sur la version anglaise de la 
-documentation `Possum <http://possum.readthedocs.org/en/>`_.
+documentation `Possum <http://doc.possum-software.org/en/>`_.